+++ lib/features/weather/components/city_search_button.dart
@@ -3,11 +3,11 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 
 import '../../../core/strings/city_search_button_strings.dart';
-import '../../../view_model/providers/city_name_validator_provider.dart';
-import '../../../view_model/providers/city_search_view_model_provider.dart';
-import '../../../view_model/providers/city_weather_notifier_provider.dart';
-import '../../../view_model/providers/error_view_model_provider.dart';
-import '../../../view_model/providers/text_editing_controller_provider.dart';
+import '../view_model/providers/city_name_validator_provider.dart';
+import '../view_model/providers/city_search_view_model_provider.dart';
+import '../view_model/providers/city_weather_notifier_provider.dart';
+import '../view_model/providers/error_view_model_provider.dart';
+import '../view_model/providers/text_editing_controller_provider.dart';
 
 /// 都市名を検索するボタンウィジェット。
 ///
